Правка замечаний
