+++ src/utils.js
@@ -15,3 +15,20 @@ export const getRandomDate = () => {
 
   return date;
 };
+
+export const getRandomCommentDate = () => {
+  const targetDate = new Date();
+  const dateMem = new Date();
+
+  targetDate.setDate(targetDate.getDate() - getRandomNumer(0, 5));
+  const options = {weekday: `long`, year: `numeric`, month: `long`, day: `numeric`, hour: `numeric`, minute: `numeric`, hour12: false};
+  const diffTime = targetDate.getDate() - dateMem.getDate();
+
+  if (diffTime > 2) {
+    return targetDate.toLocaleString(`en-US`, options);
+  } else if (diffTime === 0) {
+    return `today`;
+  } else {
+    return diffTime + ` days ago`;
+  }
+};
